+++ integration-tests/tests/src/node/analytics.ts
@@ -23,66 +23,54 @@ import { expect } from "chai";
 import { collectPlatformData } from "realm/scripts/submit-analytics";
 import { readJsonSync } from "fs-extra";
 
+type Fixture = "node" | "react-native" | "electron";
+
 describe("Analytics", () => {
-  function resolvePath(fileName: string) {
-    // tests are executed in directory `environments/node`
-    return ["..", "..", "tests", "src", "node", fileName].join(path.sep);
+  function resolvePath(fixture: Fixture) {
+    return path.resolve(__dirname, "fixtures", fixture);
   }
 
   function getRealmVersion() {
-    // tests are executed in directory `environments/node`
-    const rootPath = ["..", "..", "..", "package.json"].join(path.sep);
-    const realmPackageJson = readJsonSync(rootPath);
+    const realmPath = path.resolve(__dirname, "../../../../package.json");
+    const realmPackageJson = readJsonSync(realmPath);
     return realmPackageJson["version"];
   }
 
-  it("returns the expected version", async () => {
-    const packageJson = { version: "1.2.3" };
-    const data = await collectPlatformData(packageJson);
-
-    // common to all cases
+  function expectCommon(data: Record<string, unknown>) {
     expect(data["JS Analytics Version"]).equals(2);
     expect(data.Binding).equals("javascript");
     expect(data.Language).equals("javascript");
     expect(data["Host OS Type"]).equals(os.platform());
     expect(data["Host OS Version"]).equals(os.release());
     expect(data["Node.js version"]).equals(process.version);
+    expect(data["Realm Version"]).equals(getRealmVersion());
     expect(data.token).equals("ce0fac19508f6c8f20066d345d360fd0");
-
-    // specific to package.json
-    expect(data.Version).equals("1.2.3");
-  });
+  }
 
   it("parses node.js package.json", async () => {
-    const packageJson = readJsonSync(resolvePath("node-package.json"));
-
-    const data = await collectPlatformData(packageJson);
-    expect(data.Version).equals("1.11.1");
+    const data = await collectPlatformData(resolvePath("node"));
+    expectCommon(data);
+    expect(data.Version).equals("1.2.3");
     expect(data.Framework).equals("node.js");
     expect(data["Framework Version"]).equals(process.version);
     expect(data["JavaScript Engine"]).equals("v8");
-    expect(data["Realm Version"]).equals(getRealmVersion());
   });
 
   it("parses electron package.json", async () => {
-    const packageJson = readJsonSync(resolvePath("electron-package.json"));
-
-    const data = await collectPlatformData(packageJson);
-    expect(data.Version).equals("11.1.1");
+    const data = await collectPlatformData(resolvePath("electron"));
+    expectCommon(data);
+    expect(data.Version).equals("1.2.3");
     expect(data.Framework).equals("electron");
-    expect(data["Framework Version"]).equals("^16.0.4");
+    expect(data["Framework Version"]).equals("1.0.1");
     expect(data["JavaScript Engine"]).equals("v8");
-    expect(data["Realm Version"]).equals(getRealmVersion());
   });
 
   it("parses rn package.json", async () => {
-    const packageJson = readJsonSync(resolvePath("rn-package.json"));
-
-    const data = await collectPlatformData(packageJson);
-    expect(data.Version).equals("0.0.1");
+    const data = await collectPlatformData(resolvePath("react-native"));
+    expectCommon(data);
+    expect(data.Version).equals("1.2.3");
     expect(data.Framework).equals("react-native");
-    expect(data["Framework Version"]).equals("0.64.2");
+    expect(data["Framework Version"]).equals("1.0.1");
     expect(data["JavaScript Engine"]).equals("unknown");
-    expect(data["Realm Version"]).equals(getRealmVersion());
   });
 });